refactor(API): Introduce FxAssertContext.
