+++ subprojects/testfx-core/src/main/java/org/testfx/api/FxAssertContext.java
@@ -1,0 +1,27 @@
+package org.testfx.api;
+
+import com.google.common.annotations.Beta;
+import org.loadui.testfx.service.finder.NodeFinder;
+
+@Beta
+public class FxAssertContext {
+
+    //---------------------------------------------------------------------------------------------
+    // PRIVATE FIELDS.
+    //---------------------------------------------------------------------------------------------
+
+    private NodeFinder nodeFinder = FxSelector.selectorContext().getNodeFinder();
+
+    //---------------------------------------------------------------------------------------------
+    // GETTER AND SETTER.
+    //---------------------------------------------------------------------------------------------
+
+    public NodeFinder getNodeFinder() {
+        return nodeFinder;
+    }
+
+    public void setNodeFinder(NodeFinder nodeFinder) {
+        this.nodeFinder = nodeFinder;
+    }
+
+}
