+++ subprojects/testfx-core/src/test/java/org/testfx/api/integration/FxAssertDemoTest.java
@@ -1,4 +1,4 @@
-package org.testfx.demo;
+package org.testfx.api.integration;
 
 import javafx.application.Application;
 import javafx.scene.Scene;
